# frntEnd/openapi.yaml
openapi: 3.1.0
info:
  title: AI Workflow Builder API
  version: 1.0.0
  description: |
    REST API for the AI Workflow Builder platform.

    ## Authentication
    All endpoints (except `/auth/*` and `/health`) require a valid JWT Bearer token
    or an API Key in the `X-API-Key` header.

    ## Tenant Isolation
    Every request is scoped to a tenant. The tenant is resolved from the JWT claim
    (`tenant_id`) or from the API key row. Cross-tenant access is impossible by design.

    ## Error Envelope
    All error responses use the standard envelope below:
    ```json
    {
      "error": {
        "code": "WORKFLOW_NOT_FOUND",
        "message": "Workflow wf_abc123 not found.",
        "request_id": "req_01J..."
      }
    }
    ```

  contact:
    name: Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.workflowplatform.io/v1
    description: Production (us-east-1)
  - url: https://api.eu.workflowplatform.io/v1
    description: EU Production (eu-west-1 — v1.2+)
  - url: http://localhost:8000/v1
    description: Local development

# ─────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT. Obtain via `/auth/login` or `/auth/token/refresh`.

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Tenant API key (prefix `wfk_`). Created in the dashboard.
        Scopes are checked per-endpoint.

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable Schemas
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: WORKFLOW_NOT_FOUND
        message:
          type: string
          example: "Workflow wf_abc123 not found."
        request_id:
          type: string
          example: req_01JRXYZ
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'

    # ── Pagination ─────────────────────────────────────────────────────────────
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        totp_code:
          type: string
          description: Required when MFA is enabled on the account.
          pattern: '^\d{6}$'
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Seconds until access_token expires.
          example: 900
        token_type:
          type: string
          enum: [Bearer]
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    OAuthCallbackRequest:
      type: object
      required: [code, state]
      properties:
        code:
          type: string
        state:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, name, tenant_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        name:
          type: string
        tenant_name:
          type: string
          description: Creates a new tenant. Use invite flow for existing tenants.
    MfaSetupResponse:
      type: object
      properties:
        qr_uri:
          type: string
          description: otpauth:// URI for QR rendering.
        secret:
          type: string
          description: Raw TOTP secret for manual entry.
    MfaVerifyRequest:
      type: object
      required: [totp_code]
      properties:
        totp_code:
          type: string
          pattern: '^\d{6}$'

    # ── User ───────────────────────────────────────────────────────────────────
...
    WorkflowDefinition:
      type: object
      description: |
        Graph definition. nodes is a dict keyed by node_id.
        edges connect output ports to input ports.
        ui_metadata carries React Flow viewport state and layout hints.
      required: [nodes, edges]
      properties:
        nodes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NodeDefinition'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeDefinition'
        ui_metadata:
          $ref: '#/components/schemas/WorkflowUIMetadata'
...
    EdgeDefinition:
      type: object
      required: [id, source_node_id, source_port, target_node_id, target_port]
      properties:
        id:
          type: string
          description: Unique edge identifier (used by React Flow as edge key).
          example: "edge_node1_output__node2_input"
        source_node_id:
          type: string
        source_port:
          type: string
          description: Output port name on source node (e.g. "output", "true", "false").
        target_node_id:
          type: string
        target_port:
          type: string
          description: Input port name on target node (e.g. "input").
...
    ConversationPhase:
      type: string
      enum: [GATHERING, CLARIFYING, FINALIZING, GENERATING, COMPLETE]
      description: |
        Lifecycle phase of a chat session:
        - GATHERING: collecting initial requirements
        - CLARIFYING: asking follow-up questions for missing fields
        - FINALIZING: all requirements gathered, ready to generate
        - GENERATING: DAG generation in progress
        - COMPLETE: workflow successfully generated and saved
    RequirementSpec:
      type: object
      description: Structured requirements extracted from the conversation.
      properties:
        goal:
          type: string
          description: High-level description of what the workflow should achieve.
        trigger_type:
          type: string
          enum: [manual, scheduled, webhook]
        trigger_config:
          type: object
          additionalProperties: true
        input_sources:
          type: array
          items: { type: string }
        processing_steps:
          type: array
          items:
            type: object
            properties:
              description: { type: string }
              suggested_node_type: { type: string }
              config_hints:
                type: object
                additionalProperties: true
        integrations:
          type: array
          items: { type: string }
        output_format:
          type: string
        constraints:
          type: object
          additionalProperties: true
    ChatMessage:
      type: object
      required: [id, role, content, ts]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        ts:
          type: string
          format: date-time
    ChatSession:
      type: object
      required: [session_id, tenant_id, phase, created_at, updated_at]
      properties:
        session_id:
          type: string
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        phase:
          $ref: '#/components/schemas/ConversationPhase'
        generated_workflow_id:
          type: string
          nullable: true
        clarification_round:
          type: integer
          default: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChatSessionDetail:
      allOf:
        - $ref: '#/components/schemas/ChatSession'
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: '#/components/schemas/ChatMessage'
            requirement_spec:
              $ref: '#/components/schemas/RequirementSpec'
...
    ClarificationBlock:
      type: object
      description: |
        Structured clarification payload returned when phase is CLARIFYING.
        Frontend renders each question using its input_type.
      required: [type, questions]
      properties:
        type:
          type: string
          enum: [clarification]
        questions:
          type: array
          items:
            $ref: '#/components/schemas/ClarificationQuestion'
          minItems: 1
          maxItems: 3
    WorkflowUpdateRequest:
      type: object
      description: Payload sent from frontend when user modifies the generated DAG.
      required: [nodes, edges]
      properties:
        workflow_id:
          type: string
          description: ID of the workflow being edited. Used for audit trail.
        updated_nodes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NodeDefinition'
          description: Full node map after edits (replaces entire nodes dict).
        updated_edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeDefinition'
          description: Full edge list after edits.
        ui_metadata:
          $ref: '#/components/schemas/WorkflowUIMetadata'
          description: Updated viewport + layout mode from frontend.
    WorkflowUpdateResponse:
      type: object
      description: Validation result returned after frontend submits workflow edits.
      required: [valid, workflow]
      properties:
        valid:
          type: boolean
        workflow:
          $ref: '#/components/schemas/WorkflowDefinition'
        validation_errors:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
              node_id: { type: string, nullable: true }
          description: Present only when valid is false.
        suggestions:
          type: array
          items: { type: string }
          description: Optional AI-generated improvement hints (e.g. "Consider adding an error handler after node_3").
    ChatMessageResponse:
      type: object
      required: [session_id, phase, reply]
      properties:
        session_id:
          type: string
        phase:
          $ref: '#/components/schemas/ConversationPhase'
        reply:
          type: string
          description: AI-generated conversational text response.
        clarification:
          $ref: '#/components/schemas/ClarificationBlock'
          nullable: true
          description: |
            Structured clarification block. Populated only when phase is CLARIFYING.
            Frontend renders each question with the appropriate input widget.
        requirement_spec:
          $ref: '#/components/schemas/RequirementSpec'
          nullable: true
          description: Latest extracted requirement spec snapshot after this message.
        workflow_preview:
          $ref: '#/components/schemas/WorkflowDefinition'
          nullable: true
          description: |
            Populated when phase transitions to GENERATING or COMPLETE.
            Includes ui_config on every node and ui_metadata for the React Flow canvas.
        workflow_id:
          type: string
          nullable: true
          description: ID of the saved workflow when phase is COMPLETE.

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable Responses
  # ───────────────────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: "Bearer token is missing or expired."
              request_id: req_01JRXYZ
    Forbidden:
      description: Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found (or not visible to this tenant).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource state conflict (e.g. duplicate slug, optimistic lock).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: Validation error — request body failed schema or business rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      operationId: auth_register
      tags: [Auth]
      summary: Create a new tenant + owner account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created. Email verification sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/login:
    post:
      operationId: auth_login
      tags: [Auth]
      summary: Email + password login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/logout:
    post:
      operationId: auth_logout
      tags: [Auth]
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Token revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/token/refresh:
    post:
      operationId: auth_token_refresh
      tags: [Auth]
      summary: Rotate refresh token → new token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify-email:
    post:
      operationId: auth_verify_email
      tags: [Auth]
      summary: Verify email address with token from email link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Email verified.
        '400':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/password/reset-request:
    post:
      operationId: auth_password_reset_request
      tags: [Auth]
      summary: Send password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '204':
          description: Reset email sent (if account exists — response is always 204 to prevent enumeration).
  /auth/password/reset:
    post:
      operationId: auth_password_reset
      tags: [Auth]
      summary: Set new password using reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password updated.
        '400':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/oauth/{provider}:
    get:
      operationId: auth_oauth_redirect
      tags: [Auth]
      summary: Initiate OAuth2 flow (redirects to provider)
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, microsoft]
      responses:
        '302':
          description: Redirect to OAuth provider.
  /auth/oauth/{provider}/callback:
    post:
      operationId: auth_oauth_callback
      tags: [Auth]
      summary: Handle OAuth2 callback and issue tokens
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, microsoft]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthCallbackRequest'
      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mfa/setup:
    post:
      operationId: auth_mfa_setup
      tags: [Auth]
      summary: Generate TOTP secret for MFA enrollment
      responses:
        '200':
          description: TOTP provisioning details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaSetupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mfa/verify:
    post:
      operationId: auth_mfa_verify
      tags: [Auth]
      summary: Confirm TOTP code to enable MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: MFA enabled. Backup codes returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  backup_codes:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /chat/sessions:
    post:
      operationId: chat_create_session
      tags: [Chat]
      summary: Create a new chat session
      description: |
        Opens a new conversation session for natural-language workflow building.
        Returns a session_id to use in subsequent message calls.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Optional human-readable session title.
                  maxLength: 120
      responses:
        '201':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      operationId: chat_list_sessions
      tags: [Chat]
      summary: List all chat sessions for the current tenant
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Session list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatSession'
                  total: { type: integer }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chat/sessions/{session_id}:
    get:
      operationId: chat_get_session
      tags: [Chat]
      summary: Get full session including message history and extracted requirement spec
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full session detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chat/sessions/{session_id}/message:
    post:
      operationId: chat_send_message
      tags: [Chat]
      summary: Send a message and receive an AI reply with optional DAG preview
      description: |
        Core chat endpoint. The system automatically advances through phases:
        GATHERING -> CLARIFYING -> FINALIZING -> GENERATING -> COMPLETE.
        When requirements are fully gathered, workflow_preview is populated.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: User message text.
                  maxLength: 4096
      responses:
        '200':
          description: AI reply with current conversation state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Session already in COMPLETE phase.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /chat/sessions/{session_id}/generate:
    post:
      operationId: chat_force_generate
      tags: [Chat]
      summary: Force DAG generation with currently extracted requirements
      description: |
        Skips remaining clarification rounds and generates the workflow
        using best-effort defaults for unspecified optional fields.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Generated workflow definition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '422':
          description: Minimum requirements (goal + trigger) not yet captured.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /chat/sessions/{session_id}/workflow:
    put:
      operationId: chat_update_workflow
      tags: [Chat]
      summary: Submit frontend workflow edits — validates DAG and returns suggestions
      description: |
        Called when the user modifies nodes or edges in the DAG preview canvas.
        Runs cycle detection + port compatibility check. On success, persists the
        updated WorkflowDefinition and returns optional AI improvement suggestions.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdateRequest'
      responses:
        '200':
          description: Validation result with updated workflow and optional suggestions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowUpdateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'